Dorada delete metoda - Group, User
-Dorađene delete metode za grupu i korisnika
-Dodato kaskadno brisanje za grupu i korisnika
-Dodato da delete upiti budu @Transactional i @Modifying

+++ java-backend/src/main/java/com/ftn/ac/rs/svtkvt2023/controller/GroupController.java
@@ -139,7 +139,7 @@ public class GroupController {
         if (user == null)
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 
-        Long deleted = groupService.deleteGroup(Long.parseLong(id));
+        Integer deleted = groupService.deleteGroup(Long.parseLong(id));
 
         if (deleted != 0)
             return new ResponseEntity(deleted, HttpStatus.NO_CONTENT);
