+++ java-backend/src/main/java/com/ftn/ac/rs/svtkvt2023/service/UserService.java
@@ -17,5 +17,5 @@ public interface UserService {
 
     User updateUser(User user);
 
-    Long deleteUser(Long id);
+    Integer deleteUser(Long id);
 }
