+++ java-backend/src/main/java/com/ftn/ac/rs/svtkvt2023/service/impl/GroupServiceImpl.java
@@ -84,7 +84,7 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public Long deleteGroup(Long id) {
+    public Integer deleteGroup(Long id) {
         return groupRepository.deleteGroupById(id);
     }
 }
