+++ java-backend/src/main/java/com/ftn/ac/rs/svtkvt2023/service/impl/UserServiceImpl.java
@@ -75,7 +75,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Long deleteUser(Long id) {
+    public Integer deleteUser(Long id) {
         return userRepository.deleteUserById(id);
     }
 }
